refactor(dashboard): split test_storage_service.py by table domain (#1105 Phase 3 D6, develop-v2)
test_storage_service.py sat at its 1646-line ceiling with zero headroom. It now splits into five
modules, each under the 400-line target, cut on table domain:

  test_storage_service.py    307  core tables: defaults, shares/history, kv_store, db health,
                                  snapshot, audit_events
  test_storage_schema.py     361  schema migration, crash recovery, retention pruning
  test_storage_xvb.py        321  xvb_stats, warm standby, xvb_history, raffle_wins, payouts
  test_storage_telemetry.py  359  chart events, share stats, blocks, network history, disk growth
  test_storage_worker.py     378  worker_config and worker_history

The subject module mining_dashboard/service/storage_service.py is NOT split -- that is the binding
ruling for this cut -- so no name is repointed and all 28 top-level defs move BYTE-IDENTICAL with
zero declared edits. Proven per-block: each def is extracted from the base file by AST and must
appear verbatim in exactly one of the five modules and be absent from the other four. 28/28 pass,
1581 lines. The control fired in four separate modules: corrupting a def makes its row go red.

No fixture duplication. Unlike the D2-D5 cuts, which duplicated ~93 lines of shared builders,
state_manager and the autouse _isolate_db live in dashboard/tests/conftest.py -- the tests-ROOT
conftest -- which every module under dashboard/tests/ inherits. An AST pass over the base file
found zero module-level constants, builders and non-test functions, and zero cross-references
between top-level defs; a seeded control saw an injected constant and helper, so the empty result
is an instrument that fires. Duplication here is the import block alone.

TestAuditEvents deliberately stays in the retained module: docs/dev/testing-strategy.md:135 cites
test_storage_service.py::TestAuditEvents by path, and keeping it there keeps that citation true.

Per-module import blocks are narrowed by AST reference analysis, not by regex over the text -- a
regex counts sqlite3 inside a comment and storage_service inside the dotted patch target
"mining_dashboard.service.storage_service.random.random", which is a string literal and binds
nothing. Both over-matched on the first pass and ruff rejected them as F401.

Budget: the retained module falls to 307 lines, so its tsv entry is REMOVED -- the gate fails a
file at or under target that still carries a ceiling. All four new modules are under target and
correctly take no row. Confirmed with a positive control: restoring the stale row makes the gate
refuse it by name.

Docs: repo-map.md claimed dashboard's test files are "already one file per module". After this cut
one subject module has five test files, so the claim is corrected rather than left to rot.

Suite 2120 passed, collect-only 2120 node IDs, unchanged from the base. Node-ID identity is proven
on the multiset with the file-path component stripped -- moved tests keep their Class::test
identity exactly -- and that instrument was shown to fire by renaming one test and watching the
diff go non-empty.
