+++ docs/dev/repo-map.md
@@ -13,7 +13,7 @@ disagree, the tree is right; fix the map (see [`STYLE.md`](STYLE.md) § Accuracy
 | **Compose stack config** | The service graph (eleven services, nine on a default install) and the tunable config schema. | Root — release-bundle files the CLI reads by fixed name. | `docker-compose.yml`, `config.reference.json` (every key, documented), `config.minimal.json` / `config.core-keys.json` |
 | **Appliance image** | Pithead OS: the bootable image build (rootfs, RAUC A/B slots, the first-boot wizard, Quadlet units, the disk installer). | `os/` — the one extra top-level entry the appliance lane (`develop-v2`) carries that the Compose-only `develop` lane doesn't. | `os/build-image.sh` (image build), `os/rootfs/` (rootfs + Dockerfile), `os/installer/`, `os/README.md`, `os/KNOWN-ISSUES.md` |
 | **Release tooling** | Cutting a release (build → stage → smoke → promote → publish, cosign signing), and the lint/verification scripts `make lint`/`make test` shell out to. | `scripts/` — operational scripts, not part of any shipped image or the CLI itself. | `scripts/release.sh`, `scripts/release-smoke.sh`, `scripts/lint-*.sh`, `scripts/patch-coverage.sh` |
-| **Tests — tier 1 (unit)** | Decision logic & field mapping with the daemons stubbed: `pithead`'s own logic and the dashboard's. | `tests/stack/` (shell) and `dashboard/tests/` (pytest) — see [`testing-strategy.md`](testing-strategy.md) for the four-tier model this repo holds to. | `tests/stack/run.sh` is being split by domain (#1105 Phase 1 / #1252): each `test-*.sh` file beside it is a domain already out, `lib.sh` is the shared harness, and what remains in `run.sh` is still mid-split — the sandbox-coupled clusters land last. Issue #1252 holds the boundary map; the standalone `test_*.sh` files are separately-invoked suites, not split domains. Each domain's `source` stanza is guarded (#1400): a file that fails to load before its first assertion contributes no assertions, and the suite reds naming it rather than skipping the domain and still exiting 0. `dashboard/tests/{service,web,client,frontend,...}/test_*.py` and `*.test.mjs` are already one file per module. |
+| **Tests — tier 1 (unit)** | Decision logic & field mapping with the daemons stubbed: `pithead`'s own logic and the dashboard's. | `tests/stack/` (shell) and `dashboard/tests/` (pytest) — see [`testing-strategy.md`](testing-strategy.md) for the four-tier model this repo holds to. | `tests/stack/run.sh` is being split by domain (#1105 Phase 1 / #1252): each `test-*.sh` file beside it is a domain already out, `lib.sh` is the shared harness, and what remains in `run.sh` is still mid-split — the sandbox-coupled clusters land last. Issue #1252 holds the boundary map; the standalone `test_*.sh` files are separately-invoked suites, not split domains. Each domain's `source` stanza is guarded (#1400): a file that fails to load before its first assertion contributes no assertions, and the suite reds naming it rather than skipping the domain and still exiting 0. `dashboard/tests/{service,web,client,frontend,...}/test_*.py` and `*.test.mjs` are one file per module, except where a module's own tests outgrew the file budget and were split by domain in their own right: `storage_service.py` keeps one subject module but its tests are five files — `test_storage_service.py` (core tables) beside `test_storage_schema.py`, `test_storage_xvb.py`, `test_storage_telemetry.py` and `test_storage_worker.py`, cut on table domain in #1105 Phase 3. |
 | **Tests — tier 2 (contract)** | The real Monero/Tari clients against controllable fake daemons — proves the clients parse real wire formats. | `tests/integration/fakes/` | `tests/integration/fakes/test_contract.py`, `fake_monerod.py`, `fake_tari.py`, `contract/v1/` (RigForge's own wire fixtures, vendored) |
 | **Tests — tier 3 (mini-stack)** | The real dashboard + docker-control proxy against fake daemons, real containers — proves the control plane end-to-end (sync hold/release, node-down reject/readmit). | `tests/integration/mini-stack/` | `run-mini-stack.sh`, `docker-compose.fake.yml` |
 | **Tests — tier 4 (live matrix)** | What only reality proves: real sync, real TLS, real Tor onions, fault injection — against a real box. For the appliance, the KVM/hardware battery for a flashed image. | `tests/integration/` (Compose lane) and `tests/os/` (appliance lane) | `tests/integration/run.sh`, `tests/integration/scenarios.sh`; `tests/os/run.sh`, `tests/os/verify-image.sh` |
